Move turn_lane_direction helper to odin/util
